requirements.txt
transformers
torch
lightning
pytorch-nlp
huggingface_hub
numpy
emoji
NDETCStemmer-kaenova